remove errorFun from ark to instant parser
