+++ webapi/src/it/scala/org/knora/webapi/messages/StringFormatterSpec.scala
@@ -9,7 +9,6 @@ import java.time.Instant
 import java.util.UUID
 
 import dsp.errors.AssertionException
-import dsp.errors.BadRequestException
 import org.knora.webapi._
 import org.knora.webapi.messages.IriConversions._
 import org.knora.webapi.sharedtestdata.SharedOntologyTestDataADM
@@ -991,27 +990,6 @@ class StringFormatterSpec extends CoreSpec {
       val timestamp        = Instant.parse("2018-06-04T08:56:22Z")
       val arkUrl           = resourceIri.toSmartIri.fromResourceIriToArkUrl(maybeTimestamp = Some(timestamp))
       assert(arkUrl == "http://0.0.0.0:3336/ark:/72163/1/0001/cmfk1DMHRBiR4=_6HXpEFAn.20180604T085622Z")
-    }
-
-    "parse an ARK URL timestamp with a fractional part" in {
-      val timestampStr = "20180604T0856229876543Z"
-      val timestamp =
-        stringFormatter.arkTimestampToInstant(timestampStr, throw BadRequestException(s"invalid timestamp"))
-      assert(timestamp == Instant.parse("2018-06-04T08:56:22.9876543Z"))
-    }
-
-    "parse an ARK URL timestamp with a leading zero" in {
-      val timestampStr = "20180604T085622098Z"
-      val timestamp =
-        stringFormatter.arkTimestampToInstant(timestampStr, throw BadRequestException(s"invalid timestamp"))
-      assert(timestamp == Instant.parse("2018-06-04T08:56:22.098Z"))
-    }
-
-    "parse an ARK URL timestamp without a fractional part" in {
-      val timestampStr = "20180604T085622Z"
-      val timestamp =
-        stringFormatter.arkTimestampToInstant(timestampStr, throw BadRequestException(s"invalid timestamp"))
-      assert(timestamp == Instant.parse("2018-06-04T08:56:22Z"))
     }
 
   }
