+++ webapi/src/main/scala/org/knora/webapi/messages/ValuesValidator.scala
@@ -1,6 +1,8 @@
 package org.knora.webapi.messages
 
 import java.time.Instant
+import java.time.OffsetDateTime
+import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
 import java.time.temporal.TemporalAccessor
 import scala.util.Try
@@ -33,9 +35,7 @@ object ValuesValidator {
   }
 
   /**
-   * Checks that the format of a Knora date string is valid.
-   * If the pattern doesn't match (=> None), the date string is formally invalid.
-   * This is a mere formal validation, the actual validity check is done in `DateUtilV1.dateString2DateRange
+   * Checks that the format of a date string is valid.
    *
    * @param s a date string.
    * @return  [[Option]] of [[String]].
@@ -45,7 +45,7 @@ object ValuesValidator {
     val calendarSeparator: String  = ":"
     val precisionSeparator: String = "-"
 
-    // The expected format of a Knora date.
+    // The expected format of a date.
     // Calendar:YYYY[-MM[-DD]][ EE][:YYYY[-MM[-DD]][ EE]]
     // EE being the era: one of BC or AD
     val dateRegex: Regex = ("""^(GREGORIAN|JULIAN|ISLAMIC)""" +
@@ -77,5 +77,39 @@ object ValuesValidator {
   def xsdDateTimeStampToInstant(s: String): Option[Instant] = {
     val accessor: TemporalAccessor = DateTimeFormatter.ISO_OFFSET_DATE_TIME.parse(s)
     Try(Instant.from(accessor)).toOption
+  }
+
+  /**
+   * Parses a DSP ARK timestamp.
+   *
+   * @param timestampStr the string to be parsed.
+   * @return [[Option]] of [[Instant]].
+   */
+  def arkTimestampToInstant(timestampStr: String): Option[Instant] = {
+    val arkTimestampRegex: Regex = """^(\d{4})(\d{2})(\d{2})T(\d{2})(\d{2})(\d{2})(\d{1,9})?Z$""".r
+    timestampStr match {
+      case arkTimestampRegex(year, month, day, hour, minute, second, fraction) =>
+        val nanoOfSecond: Int = Option(fraction) match {
+          case None                  => 0
+          case Some(definedFraction) =>
+            // Pad the nano-of-second with trailing zeroes so it has 9 digits, then convert it to an integer.
+            definedFraction.padTo(9, '0').toInt
+        }
+        Try(
+          Instant.from(
+            OffsetDateTime.of(
+              year.toInt,
+              month.toInt,
+              day.toInt,
+              hour.toInt,
+              minute.toInt,
+              second.toInt,
+              nanoOfSecond,
+              ZoneOffset.UTC
+            )
+          )
+        ).toOption
+      case _ => None
+    }
   }
 }
